requirements.txt
fastapi
uvicorn
requests
playwright
groq
python-dotenv
pandas
numpy
beautifulsoup4
scikit-learn
matplotlib
seaborn
python-multipart